pyproject.toml
[project]
name = "hammer-systems-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry Shulga",email = "Shulga.dmitry.16.1992@yandex.by"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
